pyproject.toml
[project]
name = "nomoreagentsai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.4.2",
]
